Fix plural headword, phrase search.
Headword `BoolExpr`s are now applied to each headword, the result of
which is ORd together. Previously, AND was used, which excluded valid
results. For instance, if a plural definition or phrase line was
"watty, jink" and the search was "watt", then the entry would be
ommitted because only the first headword satisfied the search predicate.

Additionally, a test case for both definitions and phrases was added.
This caused the export change in `ClI.Parser.Types`, and the (small)
edit to ./examples/globLog.